requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
Django==3.1.6
idna==2.10
image==1.5.33
multidict==5.1.0
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
